+++ src/utils/priceHelper.ts
@@ -1,4 +1,19 @@
-import { Interval } from '@/services/stripe.service'
+import bytes from 'bytes';
+import { Interval } from '@/services/stripe.service';
+
+const CURRENCY_MAP = {
+    eur: '€',
+    usd: '$',
+};
+
+export interface PlanById {
+    priceId: string;
+    storage: string;
+    price: number;
+    currency: string;
+    currencyValue: string;
+    interval: Interval;
+}
 
 export const getMinimumPrice = (
     products: any,
@@ -12,3 +27,17 @@ export const getMinimumPrice = (
         ?(Math.floor(price * (decimalDiscount / 100) * 100) / 100).toFixed(2)
         : price.toFixed(2);
 }
+
+/**
+ * Normalizes a raw Stripe price (as returned by the payments API) into the shape
+ * the pricing components expect, for prices that are not part of the curated
+ * list returned by `stripeService.getPrices` (antivirus, partner deals, etc).
+ */
+export const normalizePlanById = (price: any): PlanById => ({
+    priceId: price.id,
+    storage: bytes(price.bytes),
+    price: price.decimalAmount,
+    currency: CURRENCY_MAP[price.currency] ?? price.currency,
+    currencyValue: price.currency,
+    interval: price.interval,
+});
